[style] remove unnecessary comment and organize code
