+++ src/main/java/com/Lubee/Lubee/calendar/service/HoneyService.java
@@ -5,12 +5,12 @@ import com.Lubee.Lubee.calendar.repository.CalendarRepository;
 import com.Lubee.Lubee.common.api.ApiResponseDto;
 import com.Lubee.Lubee.common.api.ErrorResponse;
 import com.Lubee.Lubee.common.api.ResponseUtils;
-import com.Lubee.Lubee.common.enumSet.ErrorType;
-import com.Lubee.Lubee.common.exception.RestApiException;
 import com.Lubee.Lubee.couple.domain.Couple;
 import com.Lubee.Lubee.couple.repository.CoupleRepository;
+import com.Lubee.Lubee.couple.service.CoupleService;
 import com.Lubee.Lubee.user.domain.User;
 import com.Lubee.Lubee.user.repository.UserRepository;
+import com.Lubee.Lubee.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -29,6 +29,8 @@ public class HoneyService {
     private final UserRepository userRepository;
     private final CoupleRepository coupleRepository;
     private final CalendarRepository calendarRepository;
+    private final UserService userService;
+    private final CoupleService coupleService;
 
     /**
      * <오늘의 꿀 개수 조회 (날짜 하루)>
@@ -39,10 +41,8 @@ public class HoneyService {
     @Transactional(readOnly = true)
     public ApiResponseDto<Integer> getHoneyInfoByUserAndDate(UserDetails userDetails, Date date){
 
-        final User user = userRepository.findUserByUsername(userDetails.getUsername())
-                .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_USER));
-        final Couple couple = coupleRepository.findCoupleByUser(user)
-                .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_COUPLE));
+        final User user = userService.getUser(userDetails);
+        final Couple couple = coupleService.getCoupleByUser(user);
         final Calendar calendar = calendarRepository.findByCoupleAndEventDate(couple, date);
 
         // 오늘의 허니 계산
@@ -68,10 +68,8 @@ public class HoneyService {
     @Transactional(readOnly = true)
     public ApiResponseDto<Long> getTotalHoneyByUser(UserDetails userDetails){
 
-        final User user = userRepository.findUserByUsername(userDetails.getUsername())
-                .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_USER));
-        final Couple couple = coupleRepository.findCoupleByUser(user)
-                .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_COUPLE));
+        final User user = userService.getUser(userDetails);
+        final Couple couple = coupleService.getCoupleByUser(user);
 
         return ResponseUtils.ok(
                 couple.getTotal_honey(),
@@ -88,31 +86,21 @@ public class HoneyService {
     @Transactional(readOnly = true)
     public ApiResponseDto<Integer> getMonthlyHoneyByUser(UserDetails userDetails, int year, int month) {
 
-        final User user = userRepository.findUserByUsername(userDetails.getUsername())
-                .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_USER));
-        final Couple couple = coupleRepository.findCoupleByUser(user)
-                .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_COUPLE));
+        final User user = userService.getUser(userDetails);
+        final Couple couple = coupleService.getCoupleByUser(user);
         List<Calendar> calendars = calendarRepository.findAllByCoupleAndYearAndMonth(couple, year, month);
 
-        int monthHoney = 0;
+        int monthlyHoney = 0;
         if (!calendars.isEmpty()) {     // calendar 리스트가 비어있지 않으면 월별 honey 개수 세기
             for (Calendar calendar : calendars) {
-                monthHoney += calendar.getCalendarMemories().size();     // 현재 Calendar의 CalendarMemory 개수를 추가하여 누적
+                monthlyHoney += calendar.getCalendarMemories().size();     // 현재 Calendar의 CalendarMemory 개수를 추가하여 누적
             }
         }
 
         return ResponseUtils.ok(
-                monthHoney,
+                monthlyHoney,
                 ErrorResponse.builder().status(200).message("요청 성공").build()
         );
-    }
-
-    @Transactional(readOnly = true)
-    public Long getTotalHoney(User user)
-    {
-        final Couple couple = coupleRepository.findCoupleByUser(user)
-                .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_COUPLE));
-        return couple.getTotal_honey();
     }
 
 }
