+++ src/main/java/com/Lubee/Lubee/common/logging/LoggingInterceptor.java
@@ -1,4 +1,4 @@
-package com.Lubee.Lubee.common;
+package com.Lubee.Lubee.common.logging;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
