+++ src/main/java/com/Lubee/Lubee/location/service/LocationService.java
@@ -42,7 +42,6 @@ public class LocationService {
 
         if(locationDtos.isEmpty()){
             return ResponseUtils.ok(null, ErrorResponse.builder().status(200).message("해당 키워드로 검색된 결과가 없습니다").build());
-            //throw new RestApiException(ErrorType.LOCATION_NOT_FOUND);
         }
 
         return ResponseUtils.ok(new LocationSearchResponse(locationDtos), ErrorResponse.builder().status(200).message("위치 검색 성공").build());
