+++ src/main/java/com/Lubee/Lubee/user_calendar_memory/repository/UserCalendarMemoryRepository.java
@@ -1,6 +1,5 @@
 package com.Lubee.Lubee.user_calendar_memory.repository;
 
-import com.Lubee.Lubee.calendar_memory.domain.CalendarMemory;
 import com.Lubee.Lubee.user.domain.User;
 import com.Lubee.Lubee.user_calendar_memory.domain.UserCalendarMemory;
 import org.springframework.data.jpa.repository.JpaRepository;
